Add initial stuff for OpenAPI

+++ Sources/TaskerServerLib/Configuration/Configuration.swift
@@ -10,7 +10,7 @@ import Configuration
 
 public class Configuration {
     public var serverName: String
-    public var serverPort: Int
+    public var serverPort: UInt16
     public var logFile: String
     public var connectionString: String
     public var secret: String
@@ -18,7 +18,7 @@ public class Configuration {
 
     init(manager: ConfigurationManager) {
         self.serverName = manager["serverName"] as? String ?? ""
-        self.serverPort = manager["serverPort"] as? Int ?? 0
+        self.serverPort = manager["serverPort"] as? UInt16 ?? 0
         self.logFile = manager["logFile"] as? String ?? ""
         self.connectionString = manager["connectionString"] as? String ?? ""
         self.secret = manager["secret"] as? String ?? ""
